Add scripts and vagrant file
The Vagrantfile sets up an Ubuntu box with all of our development
dependencies installed. It's nice to be able to just do a
`vagrant up` to get a virtual machine with every tool I use for
development set up and ready to use.

Also, I've been meaning to make all of these scripts for some time.