Fix vagrant builds and vim binding ambiguity
The version of Ubuntu I was targeting in the Vagrantfile
wasn't satisfactory. Update that and fix the problems caused
by not installing git from the git-go.

Also, there was an ambiguous binding collision in vim that
meant doing a particular action caused a delay. Move the search
leader to <leader>a to fix this.