Fix our custom ctrl-bindings to allow for <Space>
The problem is that :normal has some shitty behavior in vim
wherein it adds a `.` before commands that start with `1 `,
the way the documentation says to escape a space when calling
`normal`. This adds an `_` before the command. That also
probably breaks using movement commands from the bindings menu,
but that's the best we can do for now.